+++ CHANGES.md
@@ -17,13 +17,14 @@ The following changes are not yet released, but are code complete:
 * Add scrapers for Texas Supreme Court, Court of Criminal Appeals, and Courts of Appeal
 
 Features:
--
+- Add opinion type sanity check #1736
 
 Changes:
 -
 
 Fixes:
--
+- fix `tenn` miss adding opinion types #1749
+
 
 ## Current